requirements.txt
deeplake==4.0.3
demo1==1.0.0
matplotlib==3.5.2
moviepy==2.0.0
nltk==3.7
numpy==2.0.0
opencv_python==4.10.0.84
pandas==1.4.4
Pillow==9.2.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.64.1
transformers==4.46.2
wordfreq==3.1.1
pycocoevalcap==1.2